% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_parse_board_members.R
\name{board_members_to_tibble}
\alias{board_members_to_tibble}
\title{Convert board member list to tibble}
\usage{
board_members_to_tibble(members)
}
\arguments{
\item{members}{List of board member objects}
}
\value{
Tibble containing board member data
}
\description{
Transforms a list of board member objects into a tibble data frame.
}
