% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_parse_foreign_grants.R
\name{foreign_activities_to_tibble}
\alias{foreign_activities_to_tibble}
\title{Convert foreign activities list to tibble}
\usage{
foreign_activities_to_tibble(activities)
}
\arguments{
\item{activities}{List of foreign activity objects}
}
\value{
Tibble containing foreign activities data
}
\description{
Transforms a list of foreign activity objects into a tibble data frame.
}
