% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_parse_foreign_grants.R
\name{foreign_grants_to_tibble}
\alias{foreign_grants_to_tibble}
\title{Convert foreign grants list to tibble}
\usage{
foreign_grants_to_tibble(grants)
}
\arguments{
\item{grants}{List of foreign grant objects}
}
\value{
Tibble containing foreign grants data
}
\description{
Transforms a list of foreign grant objects into a tibble data frame.
}
