% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_parse_foreign_grants.R
\name{foreign_individual_grants_to_tibble}
\alias{foreign_individual_grants_to_tibble}
\title{Convert foreign individual grants list to tibble}
\usage{
foreign_individual_grants_to_tibble(grants)
}
\arguments{
\item{grants}{List of foreign individual grant objects}
}
\value{
Tibble containing foreign individual grants data
}
\description{
Transforms a list of foreign individual grant objects into a tibble data frame.
}
