% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_xml_parser.R
\name{parse_990_xml}
\alias{parse_990_xml}
\title{Parse a 990 XML file into a filing list object}
\usage{
parse_990_xml(xml_path)
}
\arguments{
\item{xml_path}{Path to XML}
}
\description{
Parse a 990 XML file into a filing list object
}
