% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_xml_parser.R
\name{peek_filing_ein}
\alias{peek_filing_ein}
\title{Peek EIN from early lines of a 990 XML file}
\usage{
peek_filing_ein(xml_path, max_lines = 5000L)
}
\arguments{
\item{xml_path}{File path}

\item{max_lines}{Max lines to read}
}
\description{
Peek EIN from early lines of a 990 XML file
}
